+++ Firestore/core/src/local/leveldb_remote_document_cache.h
@@ -55,7 +55,7 @@ class LevelDbRemoteDocumentCache : public RemoteDocumentCache {
   model::MutableDocumentMap GetAll(const model::DocumentKeySet& keys) override;
   model::MutableDocumentMap GetAll(const std::string& collection_group,
                                    const model::IndexOffset& offset,
-                                   int limit) const override;
+                                   size_t limit) const override;
   model::MutableDocumentMap GetAll(const model::ResourcePath& path,
                                    const model::IndexOffset& offset) override;
